refactor: merge test_cycle and test_closed_loop
